Develop

+++ backend/src/main/java/com/devsuperior/sds4/dsvendas/entities/Sale.java
@@ -18,8 +18,8 @@ public class Sale {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Integer visited;
-    private Integer deals;
+    private Long visited;
+    private Long deals;
     private Double amount;
     private LocalDate date;
